Revert "bgpd: Fix compilation for Debian 11 when printing uint64 values"
This reverts commit a44ef01d8357b9e2ea7405f3d76be5874e32a3a5.

The problem there was that `snprintf` was getting hit with the error
checking for `snprintfrr`.  The fix is actually wrong :(.

+++ bgpd/bgp_flowspec_util.c
@@ -336,8 +336,7 @@ int bgp_flowspec_op_decode(enum bgp_flowspec_util_nlri_t type,
 					ptr += len_written;
 				}
 			}
-			len_written = snprintf(ptr, len_string, " %" PRIu64 " ",
-					       (unsigned long long)value);
+			len_written = snprintf(ptr, len_string, " %" PRIu64 " ", value);
 			if (len_written > 0 && len_written < len_string) {
 				len_string -= len_written;
 				ptr += len_written;
@@ -474,8 +473,7 @@ int bgp_flowspec_bitmask_decode(enum bgp_flowspec_util_nlri_t type,
 					ptr += len_written;
 				}
 			}
-			len_written = snprintf(ptr, len_string, "%" PRIu64,
-					       (unsigned long long)value);
+			len_written = snprintf(ptr, len_string, "%" PRIu64, value);
 			if (len_written > 0 && len_written < len_string) {
 				len_string -= len_written;
 				ptr += len_written;
